backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.2.0
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
astor==0.8.1
astroid==3.2.4
async-timeout==4.0.3
attrs==25.1.0
beautifulsoup4==4.12.3
bidict==0.23.1
billiard==4.2.0
blinker==1.9.0
bs4==0.0.2
cachelib==0.13.0
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
chess==1.10.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.0
coverage==7.4.0
cycler==0.12.1
dill==0.3.8
distro==1.9.0
dnspython==2.5.0
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SocketIO==5.5.1
fonttools==4.48.1
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5validator==0.4.2
httpcore==1.0.7
httpx==0.28.1
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.5
kombu==5.4.0
langchain==0.3.17
langchain-core==0.3.33
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
langgraph==0.2.69
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.3.4
MarkupSafe==3.0.2
matplotlib==3.8.2
mccabe==0.7.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
Naked==0.1.32
numpy==1.26.4
oauthlib==3.2.2
openai==1.61.0
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.0
pillow==10.2.0
platformdirs==4.2.2
pluggy==1.5.0
priority==2.0.0
prompt_toolkit==3.0.47
propcache==0.2.1
psutil==6.0.0
pyarrow==15.0.0
pycairo==1.27.0
pycodestyle==2.12.1
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydocstyle==6.3.0
PyJWT==2.8.0
pylint==3.2.7
pymongo==4.6.1
pyparsing==3.1.1
PySocks==1.7.1
pytest==8.3.2
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2024.1
PyYAML==6.0.2
Quart==0.19.6
quart-cors==0.7.0
redis==5.0.8
redis-server==6.0.9
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
selenium==4.25.0
shellescape==3.8.1
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.37
tenacity==9.0.0
tiktoken==0.8.0
tomlkit==0.13.2
tqdm==4.67.1
trio==0.26.2
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
